backend/api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
web3==6.11.1
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0